inbox-updater/requirements.txt
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
anthropic>=0.20.0
requests>=2.28.0
python-dotenv>=1.0.0
pytz>=2023.3
